requirements.txt
amqp==2.6.1
anyjson==0.3.3
asgiref==3.3.1
astroid==2.4.2
beautifulsoup4==4.9.3
billiard==3.6.3.0
bs4==0.0.1
celery==4.4.7
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
click-default-group==1.2.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
cryptography==3.3.1
Django==3.1.5
django-rest-knox==4.1.0
djangorestframework==3.12.2
idna==2.10
isort==5.7.0
kombu==4.6.11
lazy-object-proxy==1.4.3
mccabe==0.6.1
prompt-toolkit==3.0.10
psycopg2==2.8.6
pycparser==2.20
pylint==2.6.0
python-dateutil==2.8.1
pytz==2020.5
redis==3.5.3
requests==2.25.1
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
tabulate==0.8.7
task==0.2.0
toml==0.10.2
urllib3==1.26.2
vine==1.3.0
wcwidth==0.2.5
wrapt==1.12.1

selenium~=3.141.0
pyderman~=2.0.1